+++ app/src/main/java/com.terracode.blueharvest/utils/objects/Notifications.kt
@@ -89,7 +89,7 @@ object Notifications {
 
     //Configuration Settings Notifications
 
-    fun getMaxInputDefaultNotification(context: Context, inputName: String, value: Int): Notification {
+    fun getMaxInputDefaultNotification(context: Context, inputName: String, value: Float): Notification {
         val notificationMessage = context.getString(
             R.string.maxInputNotification,
             value,
@@ -119,7 +119,7 @@ object Notifications {
         )
     }
 
-    fun getMaxInputHeightNotification(context: Context, inputName: String, value: Int): Notification {
+    fun getMaxInputHeightNotification(context: Context, inputName: String, value: Float): Notification {
         val notificationMessage = context.getString(
             R.string.maxInputNotification,
             value,
@@ -149,7 +149,7 @@ object Notifications {
         )
     }
 
-    fun safetyValueGreaterThanDisplayValueNotification(context: Context, inputName: String, value: Int): Notification {
+    fun safetyValueGreaterThanDisplayValueNotification(context: Context, inputName: String, value: Float): Notification {
         val notificationMessage = context.getString(
             R.string.safetyValueGreaterThanDisplayValueNotification,
             value,
@@ -191,7 +191,7 @@ object Notifications {
         )
     }
 
-    fun inputBelowFiveNotification(context: Context, inputName: String, value: Int): Notification {
+    fun inputBelowFiveNotification(context: Context, inputName: String, value: Float): Notification {
         val notificationMessage = context.getString(
             R.string.inputBelowFiveNotification,
             value,
@@ -205,7 +205,7 @@ object Notifications {
         )
     }
 
-    fun displayedValueLessThanSafetyValueNotification(context: Context, inputName: String, value: Int): Notification {
+    fun displayedValueLessThanSafetyValueNotification(context: Context, inputName: String, value: Float): Notification {
         val notificationMessage = context.getString(
             R.string.displayedValueLessThanSafetyValueNotification,
             value,
